pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["Farid <mustafin-farid@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "0.7.3"
hydra-core = "1.3.2"
aiogram = "3.20.0.post0"
sqlalchemy = "2.0.40"
asyncpg = "0.30.0"
lightrag-hku = "1.3.7"
aiohttp = "3.11.18"
docling = "2.35.0"
rank-bm25 = "0.2.2"
langchain = "^0.3.25"
langgraph = "^0.4.8"
langchain-openai = "^0.3.19"
langchain-community = "^0.3.24"
langfuse = "^2.60.8"

[tool.poetry.group.dev.dependencies]
pre-commit = "4.2.0"
black = "25.1.0"
isort = "6.0.1"
flake8 = "7.2.0"
ipykernel = "6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
